requirements.txt
pydantic==2.9.2
torch==2.2.2
torchvision==0.17.2
jinja2==3.1.4
fastapi==0.115.2
uvicorn==0.32.0
pillow==11.0.0
